requirements.txt
asgiref==3.2.7
boto3==1.12.38
botocore==1.15.38
Django==3.0.5
django-storages==1.9.1
docutils==0.15.2
jmespath==0.9.5
mysqlclient==1.4.6
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
